trace: cheri: function graph tracing
Support function graph tracing for cheri.

Function graph tracing adds hooks to the start and to the exit of a traced
function. Here's a basic summary of the mechanism:

start of the traced function
   jump into trampoline (ftrace_caller, arch/riscv/kernel/mcount-dyn.S)
      jump into graph handler
         save traced function's ra
         overwrite ra it with graph return handler
         call graph "start" functions
   go back to traced function

traced function exits
   "return" into graph return handler
   call graph "end" functions
   restore the original ra
   return to caller of the traced function

The graph handler (ftrace_graph_func) saves ra in struct ftrace_ret_stack's ret
component. We have to save and restore the complete register, ret must be
converted to a capability.

While the traced function runs, ra is replaced with a capability for the graph
return handler (return_to_handler) that allows executing this function.

Each task has an array of ftrace_ret_stack entries. current->ret_stack is a
pointer to this "return stack", it must become a capability.

Many places throughout the tracing code assume that a struct ftrace_ret_stack
can be saved at any current->ret_stack[i] position of the return stack. We have
to bite the bullet and make the return stack an array of uintptr_t. This wastes
some stack space when standard data fields are stored onto the stack. On the
other hand, the entries are short-lived, not passed to user space and the
recursion depth for function tracing is usually not too deep (or the trace will
become unreadable). Therefore, capability-aligned entries should be ok.

+++ arch/riscv/include/asm/ptrace.h
@@ -138,7 +138,7 @@ extern int regs_query_register_offset(const char *name);
 extern unsigned long regs_get_kernel_stack_nth(struct pt_regs *regs,
 					       unsigned int n);
 
-void prepare_ftrace_return(unsigned long *parent, unsigned long self_addr,
+void prepare_ftrace_return(uintptr_t *parent, unsigned long self_addr,
 			   unsigned long frame_pointer);
 
 /**
